Organize folder structure Fixes #9
