+++ tests/test_fastapi_app.py
@@ -1,9 +1,10 @@
 import random
 import unittest
 
-from constants import CLASSIC_FIZZBUZZ_PREFIX, N_ARGS_NAME, SEP_ARGS_NAME, STREAM_PATH
 from fastapi.testclient import TestClient
 from fastapi_app import app
+from fastapi_app.constants import (CLASSIC_FIZZBUZZ_PREFIX, N_ARGS_NAME,
+                                   SEP_ARGS_NAME, STREAM_PATH)
 from fizzbuzz_lib import classic_fizzbuzz_as_text
